added companion object for order status and fixed return issues

+++ app/src/main/java/com/live/emmazone/response_model/OrderItem.kt
@@ -7,5 +7,5 @@ data class OrderItem(
     @SerializedName("id")
     var id : Int ,
     @SerializedName("registerCode")
-    var registerCode : Int,
+    var registerCode : Long,
 ):Serializable
